Refactoring to avoid shadowing
Because the base calendar controller class from kimai uses private
methods and properties, we previously shadowed that class in this
plugin. This commit refactors the code by storing the user preferences
in a property, and overwriting the parameters that get passed on to the
view by overriding the `render` method.